Create Updated_main.py
I did all my work in Jupyter notebook, everything here should be kosher. Commented out stuff that may be extraneous when run in something like VS Code. This should result in more accurate maps and data than the previous code, at least it did for me. Also updated how some of the numbers were displayed and the formatting of the hover template for the maps.